requirements.txt
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.30.0
requests>=2.31.0